backend/etl_elastic/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Nikita <zigman.nikita@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pydantic = "^2.5.3 "
psycopg2-binary = "^2.9.9"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
urllib3 = "2.1.0"
elasticsearch = "^8.11.1"
redis = "^5.0.1"
backoff = "^2.2.1"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
ipython = "^8.13.2"
mypy = "^1.3.0"
pre-commit = "^3.3.2"
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.3.0"
types-redis = "^4.5.5.2"
ipykernel = "^6.23.1"
bandit = { extras = ["toml"], version = "^1.7.5" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
